Refactor languageservicecontext to WorkerContext
- Bundle cdn, moduleresolver and languageservicecontext under a
  WorkerContext
- Rename TypeScriptWorker to Worker/WebWorker
- Acquire Web Worker Context with only the config
- Move processFiles to typescript-builder.ts for clarity
